// src/app/Model/SearchForm.model.ts
import { SearchFormControl } from "./SearchFormControl.model";

export class SearchForm {
    departure: SearchFormControl;
    destination: SearchFormControl;
    derpartDate: SearchFormControl;
    returnDate: SearchFormControl;
    travellers: SearchFormControl;
    travelClass: SearchFormControl;

    // constructor(){
    //     this.departure = new SearchFormControl();
    //     this.destination = new SearchFormControl();
    //     this.derpartDate = new SearchFormControl();
    //     this.returnDate = new SearchFormControl();
    //     this.travelClass = new SearchFormControl();
    //     this.travellers = new SearchFormControl();
    // }
}